refactor(docs): expose agent connector slugs via Docusaurus plugin
Replaces the prebuild script that wrote src/data/agent_connectors.json
on disk with a Docusaurus plugin that discovers the connector directories
during loadContent and exposes them via setGlobalData. The
AgentConnectorRegistry component now reads the slug list with
usePluginData instead of importing a JSON file.

Removes an entire class of build footgun: the old pattern wrote the
manifest during prebuild and deleted it during postbuild, which meant
any invocation that bypassed those hooks (pnpm build, IDE integrations,
CI flows) failed with 'Can't resolve "../data/agent_connectors.json"'.

- Delete src/scripts/prepare-agent-connector-manifest.js
- Remove AGENT_CONNECTOR_MANIFEST_PATH from constants.js and cleanup-cache.js
- Drop prepare-agent-connector-manifest from prestart/prebuild
- Add src/plugins/agentConnectors.js following the same pattern as
  enterpriseConnectors.js
