+++ docusaurus/src/scripts/cleanup-cache.js
@@ -4,7 +4,7 @@
  * committed to git since it's regenerated on each build.
  */
 const fs = require("fs");
-const { REGISTRY_CACHE_PATH, AGENT_CONNECTOR_MANIFEST_PATH } = require("./constants");
+const { REGISTRY_CACHE_PATH } = require("./constants");
 
 function cleanupFile(filePath) {
   if (fs.existsSync(filePath)) {
@@ -19,7 +19,6 @@ function cleanupFile(filePath) {
 
 function cleanupCache() {
   cleanupFile(REGISTRY_CACHE_PATH);
-  cleanupFile(AGENT_CONNECTOR_MANIFEST_PATH);
 }
 
 cleanupCache();
